Move Files
1. work_extractGame
2. work_wikiAssistant
3. work_otherWiki
4. work_githubAction
