+++ work_extractGame/__init__.py
@@ -1,0 +1,3 @@
+"""
+缺氧游戏数据提取
+"""
