+++ work_githubAction/__init__.py
@@ -1,0 +1,3 @@
+"""
+定期巡查机器人脚本。挂载至GitHub Action。
+"""
