+++ work_githubAction/bot_imcopy_all.py
