+++ work_githubAction/bot_trans_img.py
