+++ work_otherWiki/__init__.py
@@ -1,0 +1,3 @@
+"""
+非缺氧wiki的脚本
+"""
