+++ work_wikiAssistant/__init__.py
@@ -1,0 +1,3 @@
+"""
+wiki站点小助手
+"""
