+++ work_wikiAssistant/bot_auto_license.py
